+++ src/snowcli/cli/snowpark/commands.py
@@ -7,6 +7,7 @@ from pathlib import Path
 from tempfile import TemporaryDirectory
 from typing import Dict, List, Set
 
+import snowcli.utils.file_utils
 import typer
 from click import ClickException
 from snowcli.cli.common.cli_global_context import cli_context
@@ -46,7 +47,7 @@ from snowcli.output.types import (
     SingleQueryResult,
 )
 from snowcli.utils import utils
-from snowcli.utils.utils import get_snowflake_packages
+from snowcli.utils.package_utils import get_snowflake_packages
 from snowcli.utils.zipper import add_file_to_existing_zip
 from snowflake.connector import DictCursor, ProgrammingError
 
@@ -447,7 +448,7 @@ def _replace_handler_in_zip(
     handler_module, _, handler_function = handler.rpartition(".")
     with TemporaryDirectory() as temp_dir:
         wrapper_file = os.path.join(temp_dir, "snowpark_coverage.py")
-        utils.generate_snowpark_coverage_wrapper(
+        snowcli.utils.file_utils.generate_snowpark_coverage_wrapper(
             target_file=wrapper_file,
             proc_name=proc_name,
             proc_signature=proc_signature,
