+++ tests/snowpark/test_snowpark_shared.py
@@ -2,13 +2,13 @@ from zipfile import ZipFile
 
 import snowcli.cli.snowpark_shared as shared
 from requirements.requirement import Requirement
-from snowcli.utils.utils import SplitRequirements
+from snowcli.utils.package_utils import SplitRequirements
 
 from tests.testing_utils.fixtures import *
 
 
-@mock.patch("tests.snowpark.test_snowpark_shared.shared.utils.parse_anaconda_packages")
-@mock.patch("tests.snowpark.test_snowpark_shared.shared.utils.install_packages")
+@mock.patch("snowcli.utils.package_utils.parse_anaconda_packages")
+@mock.patch("snowcli.utils.package_utils.install_packages")
 def test_snowpark_package(
     mock_install, mock_parse, temp_dir, correct_requirements_txt, dot_packages_directory
 ):
